refactor: extract findRoot() into separate module
Because we'll want to use the same logic from "format.js" so that we can
find the ".prettierignore" file.
